refactor: move EnrolmentButtonCell to a new file and refactor it
